requirements.txt
pillow
pillow-heif

